a working script to animate stage motion
This script is working to animate the stage motion. Four points will be plot in a box. By rotating the encoder, one of them will move accordingly. Four buttons at the outside of the encoder (left, up, right, down) can enable axis for x, y, z and rotation, switching between different axis. The button in the center could be used to switch the fine/coarse movements of the points.